retail_analytics_system/requirements.txt
opencv-python
mediapipe
numpy
ultralytics
pyyaml
flask